Security server token settings uptaded and SubUser controller config, update method added.

+++ Main-Server/Main-Server/Models/Movie.cs
@@ -18,7 +18,7 @@
 
         public string Category { get; set; }
 
-        public decimal Views { get; set; }
+        public int Views { get; set; }
 
        
     }
